backend self-destruction

+++ backend-tuucho/setup/entrypoint/entrypoint.bash
@@ -12,7 +12,19 @@ NESTJS_TIMEOUT="${NESTJS_TIMEOUT:-300}"
 # === FUNCTIONS ===
 source /usr/local/bin/bash/function/log.bash
 source /usr/local/bin/bash/function/set_key_value.bash
+source /usr/local/bin/bash/function/get_key_value.bash
 source /usr/local/bin/bash/function/wait_until.bash
+
+initialize_is_first_start() {
+    IS_FIRST_START="$(get_key_value "$IS_FIRST_START_KEY")"
+
+    if [ -z "$IS_FIRST_START" ]; then
+        IS_FIRST_START="true"
+        set_key_value "$IS_FIRST_START_KEY" "false"
+    else
+        IS_FIRST_START="false"
+    fi
+}
 
 clone_and_install() {
     log_info "Cloning and installing repo"
@@ -40,25 +52,34 @@ start_nestjs() {
 log_info "Container IP: $HOST_IP"
 
 set_key_value "$COMMAND_RESPONSE_READY_KEY" "false"
-mkdir -p "$LOG_DIR" || true
 
-if [ ! -d "$USER_HOME/$REPO_DIR/.git" ]; then
-    log_info "repository not found"
-    clone_and_install
-else
-    cd "$USER_HOME/$REPO_DIR"
-    log_info "Checking repo state"
-    git fetch origin "$BRANCH"
-    LOCAL=$(git rev-parse "$BRANCH")
-    REMOTE=$(git rev-parse "origin/$BRANCH")
-    if [ "$LOCAL" != "$REMOTE" ]; then
-        log_info "repository changed — resetting"
-        rm -rf "$USER_HOME/$REPO_DIR"
-        clone_and_install
+initialize_is_first_start
+if [ "$IS_FIRST_START" = "false" ]; then
+    CONTAINER_NAME="$(hostname)"
+    RESPONSE=$(echo "docker stop-and-delete $CONTAINER_NAME" | socat - TCP:docker-host:4777)
+    if [ "$RESPONSE" != "true" ]; then
+        log_error "$CONTAINER_NAME failed to request its own deletion"
     fi
-fi
+else
+    mkdir -p "$LOG_DIR" || true
 
-setup_receiver_socat
-start_nestjs
-set_key_value "$COMMAND_RESPONSE_READY_KEY" "true"
-wait "$nestjs_pid"
+    if [ ! -d "$USER_HOME/$REPO_DIR/.git" ]; then
+        log_info "repository not found"
+        clone_and_install
+    else
+        cd "$USER_HOME/$REPO_DIR"
+        log_info "Checking repo state"
+        git fetch origin "$BRANCH"
+        LOCAL=$(git rev-parse "$BRANCH")
+        REMOTE=$(git rev-parse "origin/$BRANCH")
+        if [ "$LOCAL" != "$REMOTE" ]; then
+            log_info "repository changed — resetting"
+            rm -rf "$USER_HOME/$REPO_DIR"
+            clone_and_install
+        fi
+    fi
+    setup_receiver_socat
+    start_nestjs
+    set_key_value "$COMMAND_RESPONSE_READY_KEY" "true"
+    wait "$nestjs_pid"
+fi
